Move: point 도메인 분리를 위해 프레젠테이션 계층 이동
