+++ src/main/kotlin/io/ticketaka/api/point/presentation/dto/BalanceResponse.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.presentation.dto
+package io.ticketaka.api.point.presentation.dto
 
 import java.math.BigDecimal
