+++ src/main/kotlin/io/ticketaka/api/point/presentation/dto/RechargeRequest.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.presentation.dto
+package io.ticketaka.api.point.presentation.dto
 
 import io.ticketaka.api.reservation.application.dto.RechargeCommand
 import java.math.BigDecimal
